WIP ffi

+++ mac-ffi-time.rb
@@ -14,7 +14,8 @@ module MacStuff
   extend FFI::Library
   ffi_lib FFI::CURRENT_PROCESS
 
-  attach_function :mach_thread_self, [], :ulong
+  attach_function :mach_thread_self, [], :uint32
+  #attach_function :thread_info, [], :int
 end
 
 Pry.start
